--- client/js/screens/Lessons/SurveyDifficulty/index.js
@@ -1,3 +1,0 @@
-import SurveyDifficulty from "./SurveyDifficultyContainer";
-
-export default SurveyDifficultyContainer;
