--- client/js/screens/Lessons/SurveyLessons/SurveyLessonsContainer.js
@@ -1,8 +1,0 @@
-import React from "react";
-import SurveyLessons from "./SurveyLessons";
-import topics from "../../../helpers/topics";
-const SurveyLessonsContainer = () => {
-  return <SurveyLessons data={topics} />;
-};
-
-export default SurveyLessonsContainer;
