--- client/js/screens/Lessons/SurveyLessons/index.js
@@ -1,2 +1,0 @@
-import SurveyLessonContainer from "./SurveyLessonsContainer";
-export default SurveyLessonContainer;
