+++ client/js/screens/SurveyDifficulty/SurveyDifficultyContainer.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import SurveyDifficulty from "./SurveyDifficulty.js";
+
+class SurveyDifficultyContainer extends Component {
+  render() {
+    return <SurveyDifficulty navigation={this.props.navigation} />;
+  }
+}
+
+export default SurveyDifficultyContainer;
