+++ client/js/screens/SurveyDifficulty/index.js
@@ -1,0 +1,3 @@
+import SurveyDifficultyContainer from "./SurveyDifficultyContainer";
+
+export default SurveyDifficultyContainer;
