+++ client/js/screens/SurveyLessons/index.js
@@ -1,0 +1,2 @@
+import SurveyLessonsContainer from "./SurveyLessonsContainer";
+export default SurveyLessonsContainer;
